edit type to value tax and total-migration invoice

+++ database/migrations/2019_11_21_042954_create_invoices_table.php
@@ -20,8 +20,8 @@ class CreateInvoicesTable extends Migration
             $table->unsignedBigInteger('invoice_state_id');
             $table->unsignedInteger('code')->unique();
             $table->date('expiration_at');
-            $table->float('value_tax');
-            $table->float('total_value');
+            $table->decimal('value_tax');
+            $table->decimal('total_value');
             $table->timestamps();
 
             $table->foreign('collaborator_id')->references('id')->on('collaborators');
